feat: add corner radius utility and refactor layout logic
- **New Features**: Introduced `rememberDeviceCornerRadius` utility to dynamically calculate the device's physical corner radius on Android 12+ (API 31) with a fallback for older versions.
- **Refactoring**:
    - Moved `WindowLayoutType` and `calculateWindowLayoutType` from `SettingsActivity` to `LayoutUtils.kt` to centralize layout calculation logic.
    - Updated `ScalePredictiveBackAnimation` to use the new centralized `rememberDeviceCornerRadius` utility.
- **Dependencies**: Updated `miuix` library from `0.8.8-ee05b131-SNAPSHOT` to `0.9.0`.
