+++ app/src/main/java/com/rosan/installer/ui/animation/predictiveback/ScalePredictiveBackAnimation.kt
@@ -1,7 +1,5 @@
 package com.rosan.installer.ui.animation.predictiveback
 
-import android.os.Build
-import android.view.RoundedCorner
 import androidx.compose.animation.AnimatedContentTransitionScope
 import androidx.compose.animation.ContentTransform
 import androidx.compose.animation.EnterExitState
@@ -21,7 +19,6 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -29,10 +26,7 @@ import androidx.compose.ui.draw.drawWithContent
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.TransformOrigin
 import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.platform.LocalView
 import androidx.compose.ui.platform.LocalWindowInfo
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.navigation3.runtime.NavKey
 import androidx.navigation3.scene.Scene
@@ -40,49 +34,11 @@ import androidx.navigation3.ui.LocalNavAnimatedContentScope
 import androidx.navigationevent.NavigationEvent.Companion.EDGE_LEFT
 import androidx.navigationevent.NavigationEventTransitionState
 import androidx.navigationevent.NavigationEventTransitionState.InProgress
+import com.rosan.installer.ui.util.rememberDeviceCornerRadius
 
 class ScalePredictiveBackAnimation : PredictiveBackAnimationHandler {
     private var exitingPageKey by mutableStateOf<String?>(null)
     private val exitAnimatable = Animatable(0f)
-
-    /**
-     * Dynamically calculates the device's physical corner radius.
-     * * @param defaultRadius The fallback radius to use if the API is lower than 31
-     * or if the device does not report rounded corners.
-     * @return The calculated corner radius in Dp.
-     */
-    @Composable
-    fun rememberDeviceCornerRadius(defaultRadius: Dp = 16.dp): Dp {
-        val view = LocalView.current
-        val density = LocalDensity.current
-
-        return remember(view, density) {
-            // Check if the device runs on Android 12 (API 31) or higher
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-                val insets = view.rootWindowInsets
-                if (insets != null) {
-                    // Try to fetch the corner radius, prioritizing the top-left corner.
-                    // Not all devices have symmetric corners, but for UI transitions,
-                    // picking one valid corner is usually sufficient.
-                    val corner = insets.getRoundedCorner(RoundedCorner.POSITION_TOP_LEFT)
-                        ?: insets.getRoundedCorner(RoundedCorner.POSITION_TOP_RIGHT)
-                        ?: insets.getRoundedCorner(RoundedCorner.POSITION_BOTTOM_LEFT)
-                        ?: insets.getRoundedCorner(RoundedCorner.POSITION_BOTTOM_RIGHT)
-
-                    if (corner != null) {
-                        // Convert pixel radius to Dp
-                        with(density) {
-                            return@remember corner.radius.toDp()
-                        }
-                    }
-                }
-            }
-
-            // Fallback to the default value
-            defaultRadius
-        }
-    }
-
 
     override suspend fun onBackPressed(
         transitionState: NavigationEventTransitionState?,
