+++ app/src/main/java/com/rosan/installer/ui/util/LayoutUtils.kt
@@ -3,6 +3,9 @@
 package com.rosan.installer.ui.util
 
 import androidx.compose.foundation.layout.BoxWithConstraintsScope
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.platform.LocalWindowInfo
 import androidx.compose.ui.unit.dp
 
 // Define standard layout thresholds based on Material Design 3 guidelines
@@ -38,3 +41,25 @@ val BoxWithConstraintsScope.isPortraitLayout: Boolean
 // Check specifically for phone landscape (wide width, but very limited height)
 val BoxWithConstraintsScope.isPhoneLandscape: Boolean
     get() = maxWidth > maxHeight && maxHeight < UIConstants.COMPACT_HEIGHT_THRESHOLD
+
+// Define screen layout types to prevent dynamic Subcompose delays
+enum class WindowLayoutType {
+    COMPACT,
+    EXPANDED
+}
+
+// Statically calculate layout based on accurate window info
+@Composable
+fun calculateWindowLayoutType(): WindowLayoutType {
+    val containerSize = LocalWindowInfo.current.containerSize
+    val density = LocalDensity.current
+    val screenWidthDp = with(density) { containerSize.width.toDp() }
+    val screenHeightDp = with(density) { containerSize.height.toDp() }
+
+    val isDefinitelyWide = screenWidthDp > UIConstants.WIDE_SCREEN_THRESHOLD
+    val aspectRatio = screenHeightDp.value / screenWidthDp.value
+    val isWideByShape = screenWidthDp > UIConstants.MEDIUM_WIDTH_THRESHOLD &&
+            aspectRatio < UIConstants.PORTRAIT_ASPECT_RATIO_THRESHOLD
+
+    return if (isDefinitelyWide || isWideByShape) WindowLayoutType.EXPANDED else WindowLayoutType.COMPACT
+}
